renamed maintenance list style
